+++ src/components/mainPage/MainScreen/EventHeader/EventHeader.stories.tsx
