+++ src/components/mainPage/MainScreen/EventHeader/EventHeader.tsx
