+++ src/components/mainPage/MainScreen/Timer/Timer.tsx
