+++ src/pages/mainPage/MainScreen/EventSelectSection/EventSelectSection.tsx
