+++ src/pages/mainPage/MainScreen/RandomEventSection/RandomEventSection.stories.tsx
