+++ src/pages/randomEventPage/RandomQuizSection/RandomQuizSection.test.tsx
